split off calculating IBD score to separate service
