+++ src/components/SnpDescentPlots.vue
@@ -77,6 +77,7 @@
   import * as d3 from 'd3'
   import { saveSvgAsPng } from 'save-svg-as-png'
   import lineReader from '../service/lineReader'
+  import identityByDecent from '../service/identityByDecent'
 
   export default {
     name: 'snp-descent-plot',
@@ -225,25 +226,6 @@
         this.dataFile = event.target.files[0]
         this.setDisableProcess()
       },
-      compareAlleles (p1, p2) {
-        if (p1 === p2 || (p1 === 'AB' && p2 === 'BA') || (p1 === 'BA' && p2 === 'AB')) {
-          return 2
-        }
-        if ((p1 === 'AA' && p2 === 'BB') || (p1 === 'BB' && p2 === 'AA')) {
-          return 0
-        }
-        if (p1 !== 'NC' && p2 !== 'NC') {
-          const p1Allele1 = p1.charAt(0)
-          const p1Allele2 = p1.charAt(1)
-          const p2Allele1 = p1.charAt(0)
-          const p2Allele2 = p1.charAt(1)
-          if (p1Allele1 === p2Allele1 || p1Allele1 === p2Allele2 || p1Allele2 === p2Allele1 || p1Allele2 === p2Allele2) {
-            return 1
-          }
-        } else {
-          return -1
-        }
-      },
       isSelectedChromosome (columns) {
         return columns[1] === this.selectedChromosome
       },
@@ -256,7 +238,7 @@
             const index2 = this.$store.state.dataIndex[combination].gPos2
             const id1 = columns[index1]
             const id2 = columns[index2]
-            const alleleScore = this.compareAlleles(id1, id2)
+            const alleleScore = identityByDecent.computeScore(id1, id2)
             // The alleleScore is equal to the key in the counts object
             this.results[combination].counts[alleleScore] += 1
             this.results[combination].points.push([parseInt(columns[2]), alleleScore])
